model_detection_2/project/requirements.txt
torchsummary
imutils
sockjs-tornado
pybase64
prometheus_client
scikit-image
xmltodict
torch-tb-profiler
torchmetrics
tqdm